+++ example/pages/todo-page.js
@@ -1,6 +1,7 @@
 const {
 	Page,
-	layout,
+	row,
+	column,
 	textInput,
 	button,
 	list,
@@ -53,7 +54,7 @@ class TodoPage extends Page {
 	inputBar() {
 		const { state } = this.useState()
 
-		return layout({
+		return row({
 			style: ['gap_2', 'p_2', 'hug'],
 			children: [
 				textInput({ model: 'inputValue', onEnter: () => this.addItem() }),
@@ -80,12 +81,12 @@ class TodoPage extends Page {
 	}
 
 	root() {
-		return layout({
-			style: ['body', 'gap_5', 'column'],
+		return column({
+			style: ['body', 'gap_5'],
 			children: [
 				header({}),
-				layout({
-					style: { size: 9, gap: 'var:u2', direction: 'column' },
+				column({
+					style: { size: 9, gap: 'var:u2' },
 					children: [
 						this.todoList(),
 						this.inputBar(),